bugfix: parameter check at removeSubscriptionEntries

+++ eventhandler/src/main/java/eu/arrowhead/core/eventhandler/database/service/EventHandlerDBService.java
@@ -382,6 +382,11 @@ public class EventHandlerDBService {
 	public void removeSubscriptionEntries( final List<Subscription> toBeRemoved ) {
 		logger.debug( "removeSubscriptionEntries started..." );
 		
+		if ( toBeRemoved == null || toBeRemoved.isEmpty()) {
+			
+			return;
+		}
+		
 		try {
 			
 			subscriptionRepository.deleteInBatch( toBeRemoved );
